refactor: remove iron-list demo module
